+++ CHANGELOG.md
@@ -7,4 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- ## [Unreleased] -->
 
-<!-- ## [0.0.1] - 2020-08-20 -->
+## [0.0.1] - 2020-08-20
+
+- Package is created.
+- Added `MomentField`, `DateMomentField`, `DatetimeMomentField` field decorators.
+- Added `DateMomentColumn`, `DatetimeMomentColumn`, `CreateMomentColumn`, `UpdateMomentColumn` column decorators for TypeORM-based entities.
+- Added formats collection as `formats` export from main module of package.